# src/WildApricot/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Wild Apricot API for non-administrative access",
    "description" : "Wild Apricot API for non-administrative access",
    "version" : "8.10.0"
  },
  "servers" : [ {
    "url" : "https://api.wildapricot.org/publicview/v1"
  } ],
  "security" : [ {
    "OAuth2_autorizationCode" : [ ]
  }, {
    "OAuth2_password" : [ ]
  } ],
  "paths" : {
...
    "/accounts/{accountId}/attachments/{attachmentId}" : {
      "get" : {
        "tags" : [ "Attachments" ],
        "summary" : "Download attachment",
        "description" : "Attachment, available for specific account.",
        "operationId" : "GetAttachmentContent",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "description" : "Your account identifier",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "format" : "integer"
          }
        }, {
          "name" : "attachmentId",
          "in" : "path",
          "description" : "Attachment identifier",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "Original", "Small", "Medium", "Medium1", "Large" ]
          }
        }, {
          "name" : "fieldSystemCode",
          "in" : "query",
          "description" : "SystemCode of attachement field which the attachment belongs to.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "contactId",
          "in" : "query",
          "description" : "ID of contact which the attachment belongs to.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventRegistrationId",
          "in" : "query",
          "description" : "ID of event registration which the attachment belongs to.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "asBase64",
          "in" : "query",
          "description" : "Base64 encoded image will be returned.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Response is common HttpResponseMessage with content set to attachment binary data.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary",
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "400" : {
            "description" : "On invalid parameters. See error details in response body.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "oAuth token was not provided, invalid or does not provide access to requested URL."
          },
          "404" : {
            "description" : "Requested item not found."
          },
          "429" : {
            "description" : "On too many requests from same account. Wait for a minute and try again."
          }
        }
      }
    },
    "/accounts/{accountId}/attachments/GetInfos" : {
      "get" : {
        "tags" : [ "Attachments" ],
        "summary" : "Get attachments infos belongs to a field value",
        "description" : "Attachment infos, available for specific account.",
        "operationId" : "GetAttachmentInfos",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "description" : "Your account identifier",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "format" : "integer"
          }
        }, {
          "name" : "fieldSystemCode",
          "in" : "query",
          "description" : "SystemCode of attachement field which the attachment belongs to.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "contactId",
          "in" : "query",
          "description" : "ID of contact which the attachment belongs to.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "eventRegistrationId",
          "in" : "query",
          "description" : "ID of event registration which the attachment belongs to.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "An array of attachment descriptions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FileInfo"
                  },
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FileInfo"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "On invalid parameters. See error details in response body.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "oAuth token was not provided, invalid or does not provide access to requested URL."
          },
          "404" : {
            "description" : "Requested item not found."
          },
          "429" : {
            "description" : "On too many requests from same account. Wait for a minute and try again."
          }
        }
      }
    },
    "/accounts/{accountId}/attachments/Upload" : {
      "post" : {
        "tags" : [ "Attachments" ],
        "summary" : "Upload attachments",
        "description" : "Upload attachments data and get their identifiers to use for contact values.  Attachment not used for any Contact field value will be deleted in 30 minutes.\n",
        "operationId" : "UploadAttachments",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "description" : "Your account identifier",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "number",
            "format" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AttachmentDataList"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "An array of attachment descriptions",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FileInfo"
                  },
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FileInfo"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "On invalid parameters. See error details in response body.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "oAuth token was not provided, invalid or does not provide access to requested URL."
          },
          "429" : {
            "description" : "On too many requests from same account. Wait for a minute and try again."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AttachmentDataList" : {
        "type" : "array",
        "description" : "list of uploading files",
        "items" : {
          "$ref" : "#/components/schemas/AttachmentData"
        }
      },
      "AttachmentData" : {
        "type" : "object",
        "properties" : {
          "Name" : {
            "type" : "string",
            "description" : "attachment name"
          },
          "MimeType" : {
            "type" : "string",
            "description" : "http ContentType (MIME type)"
          },
          "Data" : {
            "type" : "object",
            "description" : "uploading file data"
          }
        }
      },
      "FileInfo" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "description" : "Unique attachment identifier"
          },
          "Name" : {
            "type" : "string",
            "description" : "attachment name"
          },
          "ContentType" : {
            "type" : "string",
            "description" : "http ContentType (MIME type)"
          },
          "Size" : {
            "type" : "integer",
            "description" : "File size"
          },
          "CreatedDate" : {
            "type" : "string",
            "description" : "Date and time when the file has been uploaded.",
            "format" : "datetime"
          }
        },
        "example" : {
          "ContentType" : "ContentType",
          "Size" : 6,
          "CreatedDate" : "CreatedDate",
          "Id" : 0,
          "Name" : "Name"
        }
      },
...
      "Error" : {
        "required" : [ "code", "message" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Error code, like Validation / Search / ..."
          },
          "message" : {
            "type" : "string",
            "description" : "An explanation of why this error occured"
          },
          "details" : {
            "type" : "object",
            "description" : "More information about the error. Data structure is specific for each error kind."
          }
        }
      }
    },
...
    "securitySchemes" : {
      "OAuth2_autorizationCode" : {
        "type" : "oauth2",
        "description" : "\nUse this flow to get token by authorization through your organization web site login form.\n\n - Replace YOUR_ORGANIZATION to valid domain name.\n - Allow SSO for your authorized application.\n - Add your web application redirect url to list of trusted domains for that application.\n\n See also `http://gethelp.wildapricot.com/en/articles/200`\n",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://YOUR_ORGANIZATION.wildapricot.org/sys/login",
            "tokenUrl" : "https://oauth.wildapricot.org/auth/token",
            "scopes" : {
              "auto" : "autodetect allowed scopes based on current user permissions"
            }
          }
        }
      },
      "OAuth2_password" : {
        "type" : "oauth2",
        "description" : "Use this authentication flow to get oauth token by providing user credentials.\nclientId and clientSecret should be taken from Authorized applications page in admin backend.\n\nFor example if your if your emails is admin@yourdomain.com and password is 123456.\nyou should pass\n`username=admin@yourdomain.com&password=12345&&scope=auto`\n",
        "flows" : {
          "password" : {
            "tokenUrl" : "https://oauth.wildapricot.org/auth/token",
            "scopes" : {
              "auto" : "autodetect allowed scopes based on current user permissions"
            }
          }
        }
      }
    }
  }
}